streams: move ProcessorStream type off the legacy processor-runner module
The class model and browser store imported it from processor-runner.ts, which
is scheduled for deletion with the legacy runner. It now lives in types.ts;
processor-runner re-exports it until it goes.
