+++ packages/streams/src/types.ts
@@ -22,6 +22,12 @@ export type StreamEventBatch = {
 
 export type ProcessEventBatch = (batch: StreamEventBatch) => unknown;
 
+/** The minimal append surface a processor's iterate context exposes. */
+export type ProcessorStream = {
+  append(args: { streamPath?: string; event: StreamEventInput }): unknown;
+  appendBatch(args: { streamPath?: string; events: StreamEventInput[] }): unknown;
+};
+
 export type StreamRpc = {
   append(args: { streamPath?: string; event: StreamEventInput }): MaybePromise<StreamEvent>;
   appendBatch(args: {
